gitleague-client/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitleague-client"
version = "0.1.0"
description = "Push commits from local repositories to GitLeague"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "GitLeague Team", email = "hello@thegitleague.com"},
]
license = {text = "MIT"}
keywords = ["git", "league", "gamification"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "click>=8.1.0",
    "GitPython>=3.1.40",
    "httpx>=0.26.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
gitleague-client = "gitleague_client.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["gitleague_client*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
strict = true
